+++ sdk/tables/azure-data-tables/azure/data/tables/_encoder.py
@@ -12,6 +12,7 @@ from math import isnan
 
 from ._entity import EdmType, TableEntity
 from ._common_conversion import _encode_base64, _to_utc_datetime
+from ._constants import MAX_INT32, MIN_INT32, MAX_INT64, MIN_INT64
 
 _ODATA_SUFFIX = "@odata.type"
 T = TypeVar("T")
@@ -49,7 +50,12 @@ class TableEntityEncoderABC(abc.ABC, Generic[T]):
         if isinstance(value, str):
             return None, value
         if isinstance(value, int):
-            return None, value  # TODO: Test what happens if the supplied value exceeds int32.
+            if MIN_INT32 <= value <= MAX_INT32:
+                return None, value
+            elif MIN_INT64 <= value <= MAX_INT64:
+                return EdmType.INT64, str(value)
+            else:
+                return None, value
         if isinstance(value, float):
             if isnan(value):
                 return EdmType.DOUBLE, "NaN"
@@ -164,14 +170,15 @@ class TableEntityEncoder(TableEntityEncoderABC[Union[TableEntity, Mapping[str, A
         for key, value in entity.items():
             edm_type, value = self.prepare_value(key, value)
             try:
-                if _ODATA_SUFFIX in key or f"{key}{_ODATA_SUFFIX}" in entity:
+                odata = f"{key}{_ODATA_SUFFIX}"
+                if _ODATA_SUFFIX in key or odata in entity:
                     encoded[key] = value
                     continue
                 # The edm type is decided by value
                 # For example, when value=EntityProperty(str(uuid.uuid4), "Edm.Guid"),
                 # the type is string instead of Guid after encoded
                 if edm_type:
-                    encoded[f"{key}{_ODATA_SUFFIX}"] = edm_type.value if hasattr(edm_type, "value") else edm_type
+                    encoded[odata] = edm_type.value if hasattr(edm_type, "value") else edm_type
             except TypeError:
                 pass
             encoded[key] = value